changing the floating menu to support internal
Split FloatingMenu into a Server Component that picks the links and a
client shell that owns the collapse state, so entries flagged isInternal
(optionally narrowed by role) can be gated by <Internal>. Since the links
are now server-rendered, the per-link tabIndex is replaced by inert on the
<nav>, which also subsumes aria-hidden.

app/page.tsx and app/info/add/page.tsx were client components importing
FloatingMenu, which would drag next/headers and server-only into the
browser bundle. Their bodies move to HomeClient / AddInfoClient and the
pages become thin Server Components rendering the menu as a sibling. The
menu is position: fixed, so this is visually a no-op.
